requirements.txt
playwright
requests
chromadb
tqdm
python-dotenv
